Renamed FurtherDetails into SubmissionDetails for improved clarity Separated tests for each model and related forms in order to improve test readability
