+++ src/opt_out/public_api/api/admin.py
@@ -1,8 +1,7 @@
 from django.contrib import admin
 from opt_out.public_api.api.models import Submission
 
-from opt_out.public_api.api.models import FurtherDetails
+from opt_out.public_api.api.models import SubmissionDetails
 
 admin.site.register(Submission)
-admin.site.register(FurtherDetails)
-
+admin.site.register(SubmissionDetails)
